requirements.txt
asgiref==3.5.2
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
binance-connector==1.18.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cloudinary==1.31.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.4
defusedxml==0.7.1
dj-rest-auth==2.1.4
Django==3.2
django-allauth==0.50.0
django-cors-headers==3.12.0
django-environ==0.9.0
django-filter==23.2
django-rest-framework==0.1.0
django-rest-knox==4.2.0
djangorestframework==3.14.0
drf-spectacular==0.25.1
drf-standardized-errors==0.12.5
drf-yasg==1.21.4
EasyProcess==1.1
entrypoint2==1.1
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
itypes==1.2.0
jeepney==0.8.0
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.1
MouseInfo==0.1.3
mss==9.0.1
oauthlib==3.2.2
packaging==22.0
Pillow==9.5.0
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
pycparser==2.21
PyGetWindow==0.0.9
PyJWT==2.6.0
PyMsgBox==1.0.9
pyOpenSSL==23.1.1
pyperclip==1.8.2
pypng==0.20220715.0
PyRect==0.2.0
pyrsistent==0.19.3
pyscreenshot==3.1
PyScreeze==0.1.29
python3-openid==3.2.0
python3-xlib==0.15
pytweening==1.0.7
pytz==2022.6
PyYAML==6.0
qrcode==7.4.2
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
schedule==1.2.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.6.0
uritemplate==4.1.1
urllib3==1.26.13
zope.interface==6.0
